move everything to one folder
