--- src/math/operations-new/opPairUtils.ts
@@ -1,10 +1,0 @@
-import { Facet } from "data/specs/Classical"
-import { Twist } from "types"
-
-export interface FacetOpts {
-  facet?: Facet
-}
-
-export interface TwistOpts {
-  twist?: Twist
-}
