+++ src/math/operations/OperationPair.ts
